requirements.txt
pip==20.1.1
PyYAML==5.2
psycopg2-binary==2.8.4
apscheduler==3.6.3
python-dateutil==2.8.1
Flask==1.1.2
pandas==0.24.1
numpy==1.16.2
ccxt==1.28.90
gevent==20.6.0
greenlet==0.4.16
